pyproject.toml
[project]
name = "capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a,eval]>=1.18.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "python-a2a[all]>=0.5.10",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
